+++ spec/Vrac/store.spec.js
@@ -4,9 +4,23 @@ import axios from 'axios';
 import fs from 'fs';
 import path from 'path';
 
-import Vrac, { BaseModel } from '~/index';
+import { Vrac } from '~/index';
 
 Vue.use(Vuex);
+
+/**
+ * Example BaseModel.
+ */
+class BaseModel {
+    /**
+     * Setup the model.
+     *
+     * @param {object} fields
+     */
+    constructor(fields) {
+        Object.assign(this, fields);
+    }
+}
 
 describe('store', () => {
     let vrac;
@@ -99,7 +113,7 @@ describe('store', () => {
                         children: {
                             comments: new Vrac({
                                 baseUrl: 'http://localhost:3000/posts/:post_id/comments',
-                                Model: CommentModel,
+                                toModel: data => new CommentModel(data),
                             }),
                         },
                         customCalls: {
@@ -139,8 +153,8 @@ describe('store', () => {
                         expect(model).toEqual(responseModels);
                     });
 
-                    it('returns an instance of BaseModel', () => {
-                        expect(model).toBeInstanceOf(BaseModel);
+                    it('returns a normal object', () => {
+                        expect(model).not.toBeInstanceOf(BaseModel);
                     });
                 });
             });
@@ -172,7 +186,7 @@ describe('store', () => {
 
                     beforeEach(async () => {
                         responseModels = await new Promise((resolve, reject) => {
-                            const fileName = path.join(__dirname, 'netsells.ico');
+                            const fileName = path.join(__dirname, '../netsells.ico');
 
                             fs.readFile(fileName, (err, data) => {
                                 if (err) {