+++ src/Vra/index.js
@@ -4,23 +4,33 @@ import {
     parseSingle as parseSingleDefault,
     parseMultiple as parseMultipleDefault,
     parseBinary as parseBinaryDefault,
-
-    cacheMultiple as cacheMultipleDefault,
-    cacheSingle as cacheSingleDefault,
-    cacheDestroy as cacheDestroyDefault,
-    cacheBinary as cacheBinaryDefault,
-
-    BaseModel,
-} from './index';
+} from './parsers/index';
 
 /**
- * Vuex Rest API Cacher class.
+ * Vra class initilisation options.
+ *
+ * @typedef {object} VraOptions
+ * @property {string} baseUrl - URL of the endpoint, without the models ID.
+ * @property {Array<string>|string} only - Which actions to create for this model.
+ * @property {Array<string>|string} except - Which actions not to create for this model.
+ * @property {string} identifier - The identifier field, e.g. `id`.
+ * @property {object} children - Children for this module.
+ * @property {boolean} singleton - Whether this is a singleton endpoint or not - i.e. Only `read` and `update` calls.
+ * @property {Function} toModel - Convert API data to a model.
+ * @property {object} customCalls - Custom and extra API calls to add to this model.
+ *
+ * @property {Function} parseSingle - Function to parse a single item returned from the API.
+ * @property {Function} parseMultiple - Function to parse a multiple items returned from the API.
+ * @property {Function} parseBinary - Function to parse a binary item returned from the API.
+ */
+
+/**
+ * Vuex Rest API class.
  *
  * Generates Vuex stores based on the supplied config. The stores have actions
- * for fetching API data, mutators for caching them and getters for accessing
- * the cache.
+ * for fetching and returning the API data,.
  */
-class Vrac {
+class Vra {
     /**
      * Allows you to change the request function to something else (e.g. To add
      * authorization headers) or to use a different HTTP library entirely. The
@@ -43,7 +53,7 @@ class Vrac {
         const generated = {};
 
         Object.keys(modules).forEach(name => {
-            generated[name] = new Vrac(modules[name]).store;
+            generated[name] = new Vra(modules[name]).store;
         });
 
         return generated;
@@ -52,24 +62,7 @@ class Vrac {
     /**
      * Instantiate the class.
      *
-     * @param {object} options
-     * @param {string} options.baseUrl - URL of the endpoint, without the models ID.
-     * @param {Array<string>|string} options.only - Which actions to create for this model.
-     * @param {Array<string>|string} options.except - Which actions not to create for this model.
-     * @param {string} options.identifier - The identifier field, e.g. `id`.
-     * @param {object} options.children - Children for this module.
-     * @param {boolean} options.singleton - Whether this is a singleton endpoint or not - i.e. Only `read` and `update` calls.
-     * @param {Function} options.Model - Model class to use for the returned items.
-     * @param {object} options.customCalls - Custom and extra API calls to add to this model.
-     *
-     * @param {Function} options.parseSingle - Function to parse a single item returned from the API.
-     * @param {Function} options.parseMultiple - Function to parse a multiple items returned from the API.
-     * @param {Function} options.parseBinary - Function to parse a binary item returned from the API.
-     *
-     * @param {Function} options.cacheSingle - Function to cache a single item returned from the API.
-     * @param {Function} options.cacheMultiple - Function to cache multiple items returned from the API.
-     * @param {Function} options.cacheDestroy - Function to remove an item from the cache after a `destroy` call.
-     * @param {Function} options.cacheBinary - Function to cache cache a binary item returned from the API.
+     * @param {VraOptions} options
      */
     constructor({
         baseUrl = '/',
@@ -80,32 +73,22 @@ class Vrac {
         except = [],
         identifier = 'id',
         children = {},
-        Model = BaseModel,
+        toModel = data => data,
         customCalls = {},
 
         parseSingle = parseSingleDefault,
         parseMultiple = parseMultipleDefault,
         parseBinary = parseBinaryDefault,
-
-        cacheMultiple = cacheMultipleDefault,
-        cacheSingle = cacheSingleDefault,
-        cacheDestroy = cacheDestroyDefault,
-        cacheBinary = cacheBinaryDefault,
     } = {}) {
         this.baseUrl = baseUrl;
         this.identifier = identifier;
         this.calls = [];
         this.children = {};
-        this.Model = Model;
+        this.toModel = toModel;
 
         this.parseSingle = parseSingle;
         this.parseMultiple = parseMultiple;
         this.parseBinary = parseBinary;
-
-        this.cacheMultiple = cacheMultiple;
-        this.cacheSingle = cacheSingle;
-        this.cacheDestroy = cacheDestroy;
-        this.cacheBinary = cacheBinary;
 
         Object.keys(children).forEach(c => this.child(c, children[c]));
 
@@ -120,7 +103,6 @@ class Vrac {
             this.createCall('create', {
                 method: 'post',
                 parser: parseSingle,
-                cacher: cacheSingle,
                 identified: false,
             });
         }
@@ -128,9 +110,7 @@ class Vrac {
         if (includeCalls.includes('read')) {
             this.createCall('read', {
                 parser: parseSingle,
-                cacher: cacheSingle,
                 identified: !singleton,
-                readCache: !singleton,
             });
         }
 
@@ -138,7 +118,6 @@ class Vrac {
             this.createCall('update', {
                 method: 'patch',
                 parser: parseSingle,
-                cacher: cacheSingle,
                 identified: !singleton,
             });
         }
@@ -147,7 +126,6 @@ class Vrac {
             this.createCall('destroy', {
                 method: 'delete',
                 parser: parseSingle,
-                cacher: cacheDestroy,
                 identified: !singleton,
             });
         }
@@ -191,13 +169,13 @@ class Vrac {
      * Add a child module to this model.
      *
      * @param {string} name - Name of the child model.
-     * @param {object|Vrac} child - Vrac constructor options or Vrac object.
+     * @param {object|Vra} child - Vra constructor options or Vra instance.
      */
     child(name, child) {
         this.children[name]
-            = child instanceof Vrac
+            = child instanceof this.constructor
                 ? child
-                : new Vrac(child);
+                : new this.constructor(child);
     }
 
     /**
@@ -218,32 +196,13 @@ class Vrac {
     }
 
     /**
-     * Get the default cacher.
-     *
-     * @param {object} options
-     * @param {boolean} options.identified - Whether this is a cacher for an identified API call or not.
-     * @param {boolean} options.binary - Whether this is a cacher for a binary API call or not.
-     *
-     * @returns {Function}
-     */
-    getCacher({ identified, binary }) {
-        if (binary) {
-            return this.cacheBinary;
-        }
-
-        return identified ? this.cacheSingle : this.cacheMultiple;
-    }
-
-    /**
      * Create an action for an endpoint.
      *
      * @param {string} name - Name of the action.
      * @param {object} options
      * @param {string} options.method - HTTP method for the call.
      * @param {Function} options.parser - Function used to parse the data from the API response.
-     * @param {Function} options.cacher - Function used to cache the model from the response.
      * @param {boolean} options.identified - Whether this endpoint needs an identifier field or not, e.g. `id`.
-     * @param {boolean} options.readCache - Whether this action should return from the cache if the model exists there.
      * @param {string} options.path - Path for this callback, appended to baseUrl.
      * @param {boolean} options.binary - Whether this is a binary model or not.
      * @param {string} options.responseType - Override responseType, by default this is `undefined` for normal models, and `arraybuffer` for binary models.
@@ -251,24 +210,48 @@ class Vrac {
     createCall(name, {
         method = 'get',
         identified = false,
-        readCache = false,
         path = '',
         binary = false,
         responseType = binary ? 'arraybuffer' : undefined,
         parser = this.getParser({ identified, binary }),
-        cacher = this.getCacher({ identified, binary }),
+        ...rest
     } = {}) {
         this.calls.push({
             name,
             method,
             parser,
-            cacher,
             identified,
-            readCache,
             path,
             responseType,
             binary,
+            ...rest,
         });
+    }
+
+    /**
+     * Get a calls options.
+     *
+     * @param {string} name - Name of the action.
+     * @returns {object|null}
+     */
+    getCall(name) {
+        return this.calls.find(c => c.name === name);
+    }
+
+    /**
+     * Modify an action for an endpoint.
+     *
+     * @param {string} name - Name of the action.
+     * @param {object} options
+     */
+    modifyCall(name, options) {
+        const call = this.getCall(name);
+
+        if (!call) {
+            return;
+        }
+
+        Object.assign(call, options);
     }
 
     /**
@@ -285,7 +268,7 @@ class Vrac {
             return fieldsOrData;
         }
 
-        return new this.Model(fieldsOrData);
+        return this.toModel(fieldsOrData);
     }
 
     /**
@@ -304,75 +287,6 @@ class Vrac {
     }
 
     /**
-     * Get the default state for this module.
-     *
-     * @returns {object} State.
-     */
-    get state() {
-        return {
-            index: [],
-            actionsLoading: {},
-        };
-    }
-
-    /**
-     * Get the mutators for this module.
-     *
-     * @returns {object} Mutators.
-     */
-    get mutations() {
-        return {
-            createOrUpdate: (state, model) => {
-                state.index = [
-                    ...state.index.filter(m => m[this.identifier] !== model[this.identifier]),
-                    model,
-                ];
-            },
-
-            destroy: (state, model) => {
-                state.index = state.index.filter(
-                    m => m[this.identifier] !== model[this.identifier],
-                );
-            },
-
-            loading: (state, action) => {
-                state.actionsLoading = Object.assign(
-                    {},
-                    state.actionsLoading,
-                    { [action]: (state.actionsLoading[action] || 0) + 1 },
-                );
-            },
-
-            loaded: (state, action) => {
-                state.actionsLoading = Object.assign(
-                    {},
-                    state.actionsLoading,
-                    { [action]: state.actionsLoading[action] - 1 },
-                );
-            },
-        };
-    }
-
-    /**
-     * Get the getters for this module.
-     *
-     * @returns {object} Getters.
-     */
-    get getters() {
-        return {
-            index: ({ index }) => index,
-
-            read: (state, getters) => identifier => getters.index.find(
-                m => m[this.identifier] === identifier,
-            ),
-
-            loading: ({ actionsLoading }) => {
-                return Object.keys(actionsLoading).some(k => actionsLoading[k]);
-            },
-        };
-    }
-
-    /**
      * Get the actions for this module.
      *
      * @returns {object} Actions.
@@ -386,7 +300,6 @@ class Vrac {
                 fields = {},
                 params = {},
                 method = call.method,
-                readCache = call.readCache,
                 path = call.path,
                 responseType = call.responseType,
             } = {}) {
@@ -394,26 +307,8 @@ class Vrac {
                     if ([null, undefined].includes(fields[self.identifier])) {
                         throw new Error(`The '${ call.name }' action requires a 'fields.${ self.identifier }' option`);
                     }
-
-                    if (readCache) {
-                        const model = context.getters.read(fields[self.identifier]);
-
-                        if (model) {
-                            return self.createModel(model, call);
-                        }
-                    }
-                } else {
-                    if (fields[self.identifier]) {
-                        throw new Error(`The '${ call.name }' action can not be used with the 'fields.${ self.identifier }' option`);
-                    }
-
-                    if (readCache) {
-                        const cachedModels = context.getters.index;
-
-                        if (cachedModels.length) {
-                            return cachedModels.map(m => self.createModel(m, call));
-                        }
-                    }
+                } else if (fields[self.identifier]) {
+                    throw new Error(`The '${ call.name }' action can not be used with the 'fields.${ self.identifier }' option`);
                 }
 
                 let data = undefined;
@@ -422,25 +317,15 @@ class Vrac {
                     data = Object.assign({}, fields);
                 }
 
-                let response;
-
-                try {
-                    context.commit('loading', call.name);
-
-                    response = await self.constructor.requestAdapter.call(this, {
-                        url: self.getUrl(fields, path),
-                        method,
-                        data,
-                        params,
-                        responseType,
-                    }, context);
-                } finally {
-                    context.commit('loaded', call.name);
-                }
+                const response = await self.constructor.requestAdapter.call(this, {
+                    url: self.getUrl(fields, path),
+                    method,
+                    data,
+                    params,
+                    responseType,
+                }, context);
 
                 const parsed = call.parser(response.data, fields);
-
-                call.cacher(context, parsed.models || parsed.model);
 
                 if (parsed.models) {
                     const value = parsed.models.map(m => self.createModel(m, call));
@@ -463,17 +348,14 @@ class Vrac {
      * @returns {object} Store.
      */
     get store() {
-        const { actions, getters, mutations, state, modules } = this;
+        const { actions, modules } = this;
 
         return {
             namespaced: true,
             actions,
-            getters,
-            mutations,
-            state,
             modules,
         };
     }
 }
 
-export default Vrac;
+export default Vra;