+++ src/Vrac/cachers/index.js
@@ -1,0 +1,11 @@
+import cacheMultiple from './multiple';
+import cacheSingle from './single';
+import cacheDestroy from './destroy';
+import cacheBinary from './binary';
+
+export {
+    cacheMultiple,
+    cacheSingle,
+    cacheDestroy,
+    cacheBinary,
+};